requirements.txt
Django==2.2
mysqlclient==1.4.6
Jinja2==2.11.1
django-debug-toolbar==2.2
django-js-reverse==0.9.1
python-memcached==1.59
django-cache-machine==1.1.0
djangorestframework==3.11.0
djangorestframework-camel-case==1.1.2
djangorestframework-bulk==0.2.1
drf-writable-nested==0.5.4
django-webpack-loader==0.6.0
django-login-required-middleware==0.5.0
django-filter==2.2.0
imutils==0.5.3
scikit-image==0.16.2
opencv-python==4.2.0.34
django-imagefield==0.11.3

